// components/canary-ui/layout/types.ts
export enum DividerDirection {
  HORIZONTAL = "horizontal",
  VERTICAL = "vertical",
}

export enum DialogStretch {
  COMPACT = "compact",
  NORMAL = "normal",
  FILL = "fill",
}

export enum DialogCloseButtonSize {
  DEFAULT = "default",
  LARGE = "large",
  TABLET = "tablet",
}

export enum SettingsCardState {
  VIEW = "view",
  NEW = "new",
  EDIT = "edit",
  DISABLED = "disabled",
}